Redirect base domain to latest docs
The `index.html` file redirects the base URL to the `latest/` folder so
That `https://genericmappingtools.github.io/GMT.jl` goes there instead
of ending in a 404.
Add the `.nojekyll` file so that Github knows not to try to compile the
sources in the `gh-pages` branch using Jekyll (a static site generator).