add methods to `<` and `<=` instead of to `>` and `>=`
As documented, the intended way to implement `>` is to add a method to
`<`. Similarly with `>=`.

This should solve some of the problems mentioned on issue #1799.
However, for the long term (whenever a breaking release of GMT.jl
happens), I'd suggest just getting rid of the methods, as discussed on
issue #1799.