Export wkbPointZM from Gdal module
`wkbPointZM` is defined in `Gdal` and used within `ogr2GMTdataset`
with the expectation that it had been brought into scope of the
parent module with the `using Gdal` statement.  However,
`wkbPointZM` was not actually exported from `Gdal`, causing an
error when converting some files.

Add `wkbPointZM` to the list of exports in `Gdal` to fix this.